+++ src/FSharp.Formatting.CommandTool/ConvertCommand.fs
@@ -1,4 +1,4 @@
-namespace FSharp.Formatting.CommandTool.Literate
+namespace FSharp.Formatting.CommandTool
 
 open CommandLine
 open CommandLine.Text
@@ -10,7 +10,7 @@ open FSharp.Formatting.CommandTool.Common
 
 /// Process directory containing a mix of Markdown documents and F# Script files
 [<Verb("convert", HelpText = "convert a directory of literate scripts or markdown to another format")>]
-type ConvertDirectoryOptions() =
+type ConvertCommand() =
 
     // does not work as desired in F#:
     // the HelpOption attribute is not built,
@@ -21,74 +21,59 @@ type ConvertDirectoryOptions() =
         let help = new HelpText()
         help.AddDashesToOption <- true
         //help.AddOptions(x)
-        "\nfsformatting literate --processDirectory [options]" +
+        "\nfsdocs convert --processDirectory [options]" +
         "\n--------------------------------------------------" +
         help.ToString()
 
-    [<Option("help", Required = false,
-        HelpText = "Display this message. All options are case-insensitive.")>]
+    [<Option("help", Required = false, HelpText = "Display this message. All options are case-sensitive.")>]
     member val help = false with get, set
 
-    [<Option("waitForKey", Required = false,
-        HelpText = "Wait for key before exit.")>]
+    [<Option("waitForKey", Required = false, HelpText = "Wait for key before exit.")>]
     member val waitForKey = false with get, set
 
     // default settings will be mapped to 'None'
 
-    [<Option("input", Required = true,
-        HelpText = "Input directory of *.fsx and *.md files.")>]
+    [<Option("input", Required = false, Default="docs", HelpText = "Input directory of *.fsx and *.md files.")>]
     member val input = "" with get, set
 
-    [<Option("template", Required = false,
-        HelpText = "Template file for formatting (optional).")>]
+    [<Option("template", Required = false, HelpText = "Template file for formatting (optional).")>]
     member val template = "" with get, set
 
-    [<Option("output", Required = false,
-        HelpText = "Ouput Directory, defaults to 'output' (optional).")>]
+    [<Option("output", Required = false, Default="output", HelpText = "Ouput Directory, defaults to 'output' (optional).")>]
     member val output = "" with get, set
 
-    [<Option("format", Required = false,
-        HelpText = "Ouput format either 'latex', 'ipynb' or 'html', defaults to 'html' (optional).")>]
+    [<Option("format", Required = false, HelpText = "Ouput format either 'latex', 'ipynb' or 'html', defaults to 'html' (optional).")>]
     member val format = "html" with get, set
 
 //    [<Option("formatAgent", Required = false,
 //        HelpText = "FSharp Compiler selection, defaults to 'FSharp.Compiler.dll' which throws a 'file not found' exception if not in search path (optional).")>]
 //    member val fsharpCompiler = "" with get, set
 
-    [<Option("prefix", Required = false,
-        HelpText = "Prefix for formatting, defaults to 'fs' (optional).")>]
+    [<Option("prefix", Required = false, HelpText = "Prefix for formatting, defaults to 'fs' (optional).")>]
     member val prefix = "" with get, set
 
-    [<Option("compilerOptions", Required = false,
-        HelpText = "Compiler Options (optional).")>]
+    [<Option("compilerOptions", Required = false, HelpText = "Compiler Options (optional).")>]
     member val compilerOptions = Seq.empty<string> with get, set
 
-    [<Option("noLineNumbers", Required = false,
-        HelpText = "Don't add line numbers, default is to add line numbers (optional).")>]
+    [<Option("noLineNumbers", Required = false, HelpText = "Don't add line numbers, default is to add line numbers (optional).")>]
     member val noLineNumbers = false with get, set
 
-    [<Option("references", Required = false,
-        HelpText = "Turn all indirect links into references, defaults to 'false' (optional).")>]
+    [<Option("references", Required = false, HelpText = "Turn all indirect links into references, defaults to 'false' (optional).")>]
     member val references = false with get, set
 
-    [<Option("fsieval", Required = false,
-        HelpText = "Use the default FsiEvaluator, defaults to 'false'")>]
-    member val fsieval = false with set, get
+    [<Option("eval", Required = false, HelpText = "Use the default FsiEvaluator, defaults to 'false'")>]
+    member val eval = false with set, get
 
-    [<Option("noRecursive", Required = false,
-        HelpText = "Disable recursive processing of sub-directories")>]
+    [<Option("noRecursive", Required = false, HelpText = "Disable recursive processing of sub-directories")>]
     member val noRecursive = false with set, get
 
-    [<Option("parameters", Required = false,
-        HelpText = "A whitespace separated list of string pairs as text replacement patterns for the format template file (optional).")>]
+    [<Option("parameters", Required = false, HelpText = "A whitespace separated list of string pairs as text replacement patterns for the format template file (optional).")>]
     member val parameters = Seq.empty<string> with get, set
 
-    [<Option("includeSource", Required = false,
-        HelpText = "Include sourcecode in documentation, defaults to 'false' (optional).")>]
+    [<Option("includeSource", Required = false, HelpText = "Include sourcecode in documentation, defaults to 'false' (optional).")>]
     member val includeSource = false with get, set
 
-    [<Option("live", Required = false,
-        HelpText = "Watches for changes in the input directory and re-runs, if a change occures")>]
+    [<Option("live", Required = false, HelpText = "Watches for changes in the input directory and re-runs, if a change occures")>]
     member val live = false with get, set
 
     member x.Execute() =
@@ -103,7 +88,7 @@ type ConvertDirectoryOptions() =
                         x.input,
                         ?generateAnchors = Some true,
                         ?template = (evalString x.template),
-                        ?outputDirectory = Some (if x.output = "" then "output" else x.output),
+                        ?outputDirectory = Some x.output,
                         ?format=
                             Some (let fmt = x.format.ToLower()
                                   if fmt = "html" then OutputKind.Html
@@ -116,8 +101,8 @@ type ConvertDirectoryOptions() =
                         ?lineNumbers = Some (not x.noLineNumbers),
                         ?processRecursive = Some (not x.noRecursive),
                         ?references = Some x.references,
-                        ?fsiEvaluator = (if x.fsieval then Some ( FsiEvaluator() :> _) else None),
-                        ?parameters = (evalPairwiseStrings x.parameters),
+                        ?fsiEvaluator = (if x.eval then Some ( FsiEvaluator() :> _) else None),
+                        ?parameters = evalPairwiseStrings x.parameters,
                         ?includeSource = Some x.includeSource)
 
                 if x.live then